Move validation module and providers
