+++ src/domain/safe/module-transaction.validator.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@nestjs/common';
 import { ValidateFunction } from 'ajv';
-import { IValidator } from '../interfaces/validator.interface';
-import { GenericValidator } from '../schema/generic.validator';
-import { JsonSchemaService } from '../schema/json-schema.service';
-import { ModuleTransaction } from './entities/module-transaction.entity';
-import { moduleTransactionSchema } from './entities/schemas/module-transaction.schema';
+import { GenericValidator } from '../../validation/providers/generic.validator';
+import { JsonSchemaService } from '../../validation/providers/json-schema.service';
 import {
   dataDecodedParameterSchema,
   dataDecodedSchema,
 } from '../data-decoder/entities/schemas/data-decoded.schema';
+import { IValidator } from '../interfaces/validator.interface';
+import { ModuleTransaction } from './entities/module-transaction.entity';
+import { moduleTransactionSchema } from './entities/schemas/module-transaction.schema';
 
 @Injectable()
 export class ModuleTransactionValidator
