+++ src/domain/safe/safe-list.validator.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { ValidateFunction } from 'ajv';
+import { GenericValidator } from '../../validation/providers/generic.validator';
+import { JsonSchemaService } from '../../validation/providers/json-schema.service';
 import { IValidator } from '../interfaces/validator.interface';
-import { JsonSchemaService } from '../schema/json-schema.service';
-import { GenericValidator } from '../schema/generic.validator';
-import { safeListSchema } from './entities/schemas/safe-list.schema';
 import { SafeList } from './entities/safe-list.entity';
+import { safeListSchema } from './entities/schemas/safe-list.schema';
 
 @Injectable()
 export class SafeListValidator implements IValidator<SafeList> {
