+++ src/routes/delegates/pipes/create-delegate.dto.validation.pipe.ts
@@ -1,7 +1,7 @@
 import { HttpStatus, Injectable, PipeTransform } from '@nestjs/common';
 import { ValidateFunction } from 'ajv';
-import { GenericValidator } from '../../../domain/schema/generic.validator';
-import { JsonSchemaService } from '../../../domain/schema/json-schema.service';
+import { GenericValidator } from '../../../validation/providers/generic.validator';
+import { JsonSchemaService } from '../../../validation/providers/json-schema.service';
 import { CreateDelegateDto } from '../entities/create-delegate.dto.entity';
 import { createDelegateDtoSchema } from '../entities/schemas/create-delegate.dto.schema';
