+++ src/routes/delegates/pipes/get-delegate.dto.validation.pipe.ts
@@ -1,7 +1,7 @@
 import { HttpStatus, Injectable, PipeTransform } from '@nestjs/common';
 import { ValidateFunction } from 'ajv';
-import { GenericValidator } from '../../../domain/schema/generic.validator';
-import { JsonSchemaService } from '../../../domain/schema/json-schema.service';
+import { GenericValidator } from '../../../validation/providers/generic.validator';
+import { JsonSchemaService } from '../../../validation/providers/json-schema.service';
 import { GetDelegateDto } from '../entities/get-delegate.dto.entity';
 import { getDelegateDtoSchema } from '../entities/schemas/get-delegate.dto.schema';
