+++ src/validation/validation.module.ts
@@ -1,7 +1,7 @@
 import { Global, Module } from '@nestjs/common';
-import { GenericValidator } from './domain/schema/generic.validator';
-import { JsonSchemaService } from './domain/schema/json-schema.service';
-import { ValidationErrorFactory } from './domain/schema/validation-error-factory';
+import { GenericValidator } from './providers/generic.validator';
+import { JsonSchemaService } from './providers/json-schema.service';
+import { ValidationErrorFactory } from './providers/validation-error-factory';
 
 @Global()
 @Module({
